Bring healthgrades spider back online
HealthGrades has changed their site, modified the parser to reflect changes.
JSON data contains all information about the physician, so just use that as the source of truth.
Parser was not crawling out of the box, adding a catch-all rule fixes that. But we don't care about their video or static content, so included some deny rules to skip it.
Added raw JSON blob to the results, there is a lot more good info to parse.
Added TODOs to enhance readability and future HealthGrade layout changes.